requirements.txt
Flask==3.0.0
fyers-apiv3==3.1.6
pandas==2.0.3
numpy==1.24.3
APScheduler==3.10.4
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3
gunicorn==21.2.0
